Refactor supervisor detailed status assembly
